refactor: update import paths for consistency across dataset and notebook components
